# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowSentinel API",
    "description": "Metadata-only analysis of authorized packet captures. No endpoint returns packet payloads, and TLS is never decrypted. Every endpoint except POST /auth/login needs a signed-in session; viewers read, analysts also import and triage, admins also delete, change settings, manage accounts and read the audit log.",
    "contact": {
      "name": "exosphere8"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/session": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "The current session.",
        "operationId": "session",
        "responses": {
          "200": {
            "description": "The signed-in account and its CSRF token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List accounts (admin).",
        "operationId": "list_users",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "1-based page number (default 1, at most 1000000).",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "maximum": 1000000,
              "minimum": 1
            }
          },
          {
            "name": "per_page",
            "in": "path",
            "description": "Items per page, 1-500 (default 50).",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "maximum": 500,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of accounts, by username",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paged_User"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Needs the admin role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create an account (admin).",
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username or weak password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Needs the admin role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The username is taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Malformed body or unknown role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, for example `not_found`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation."
          },
          "position": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Position",
                "description": "For filter errors: the byte range `[start, end)` of the problem in\nthe filter text."
              }
            ]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "NewUser": {
        "type": "object",
        "description": "Body of `POST /users`.",
        "required": [
          "username",
          "password",
          "role"
        ],
        "properties": {
          "password": {
            "type": "string",
            "description": "12 to 256 characters, not containing the username."
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "username": {
            "type": "string",
            "description": "1 to 64 letters, digits, `.`, `_` or `-`; stored lowercase."
          }
        },
        "additionalProperties": false
      },
...
      "Paged_User": {
        "type": "object",
        "description": "A page of results.",
        "required": [
          "items",
          "page",
          "per_page",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "An account, without its password hash.",
              "required": [
                "id",
                "username",
                "role",
                "disabled",
                "created_at",
                "updated_at",
                "password_changed_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string"
                },
                "disabled": {
                  "type": "boolean",
                  "description": "A disabled account cannot sign in, and its sessions end."
                },
                "id": {
                  "type": "integer",
                  "format": "int64"
                },
                "last_login_at": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "password_changed_at": {
                  "type": "string"
                },
                "role": {
                  "$ref": "#/components/schemas/Role"
                },
                "updated_at": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              }
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total matching items across all pages."
          }
        }
      },
...
      "Position": {
        "type": "object",
        "description": "A byte range in a request value.",
        "required": [
          "start",
          "end"
        ],
        "properties": {
          "end": {
            "type": "integer",
            "minimum": 0
          },
          "start": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "Role": {
        "type": "string",
        "description": "What an account may do. Each role includes everything the roles below it\nmay do: `viewer` < `analyst` < `admin`.",
        "enum": [
          "viewer",
          "analyst",
          "admin"
        ]
      },
...
      "SessionAccount": {
        "type": "object",
        "description": "The signed-in account, as the dashboard sees it.",
        "required": [
          "id",
          "username",
          "role"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "SessionInfo": {
        "type": "object",
        "description": "A signed-in session.",
        "required": [
          "user",
          "csrf_token",
          "expires_at",
          "idle_timeout_seconds"
        ],
        "properties": {
          "csrf_token": {
            "type": "string",
            "description": "Send this in the `X-CSRF-Token` header with every `POST`, `PUT`,\n`PATCH` and `DELETE` request."
          },
          "expires_at": {
            "type": "string",
            "description": "When the session ends at the latest (RFC 3339 UTC)."
          },
          "idle_timeout_seconds": {
            "type": "integer",
            "format": "int64",
            "description": "The session also ends after this many seconds without a request.",
            "minimum": 0
          },
          "user": {
            "$ref": "#/components/schemas/SessionAccount"
          }
        }
      },
...
      "User": {
        "type": "object",
        "description": "An account, without its password hash.",
        "required": [
          "id",
          "username",
          "role",
          "disabled",
          "created_at",
          "updated_at",
          "password_changed_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "disabled": {
            "type": "boolean",
            "description": "A disabled account cannot sign in, and its sessions end."
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "last_login_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "password_changed_at": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "updated_at": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "flowsentinel_session",
        "description": "Session cookie set by POST /auth/login (named __Host-flowsentinel_session when FLOWSENTINEL_SECURE_COOKIES=true). State-changing requests must also send the session's CSRF token in the X-CSRF-Token header."
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ],
...
}